+++ torchao/__init__.py
@@ -15,15 +15,11 @@ if not _IS_FBCODE:
         logging.info("Skipping import of cpp extensions")
 
 from torchao.quantization import (
-    apply_weight_only_int8_quant,
-    apply_dynamic_quant,
     autoquant,
 )
 from . import dtypes
 
 __all__ = [
     "dtypes",
-    "apply_dynamic_quant",
-    "apply_weight_only_int8_quant",
     "autoquant",
 ]
