+++ torchao/quantization/dynamic_quant.py
@@ -7,7 +7,7 @@
 import torch
 import torch.nn as nn
 
-from .quant_primitives import (
+from .utils import (
     dynamically_quantize_per_channel,
     quant_int8_dynamic_per_token_linear,
 )
